requirements.txt
streamlit
azure-cognitiveservices-speech
dotenv